backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Vector DB
chromadb==0.4.22

# AI/ML
openai==1.12.0
google-generativeai==0.3.2

# Video Processing
yt-dlp==2024.1.1
openai-whisper==20231117

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
aiofiles==23.2.1

# Background Tasks
celery==5.3.6
redis==5.0.1
